Backend/app/worker/flower.requirements.txt
aiofiles
biopython
celery[redis]
flower
python-dotenv
redis-py
fastapi
aiohttp
minio